Prepare territoires data on import
