+++ lib/projets-validator.js
@@ -3,9 +3,7 @@
 import Joi from 'joi'
 import {parse, isValid} from 'date-fns'
 import {validatePayload} from '../server/util/payload.js'
-import {createGeometryBuilder} from '../lib/build-geometry.js'
-
-const geometryBuilder = await createGeometryBuilder()
+import {getTerritoire} from '../lib/territoires.js'
 
 function validatePerimetre(perimetre) {
   const fields = perimetre.split(':')
@@ -15,7 +13,7 @@ function validatePerimetre(perimetre) {
   }
 
   // Check if territory exist
-  geometryBuilder.getTerritoire(perimetre)
+  getTerritoire(perimetre)
 
   return perimetre
 }
